+++ demos/chinook/src/main/java/is/codion/framework/demos/chinook/domain/impl/ChinookImpl.java
@@ -308,6 +308,7 @@ public final class ChinookImpl extends DefaultDomain implements Chinook {
             Track.UNITPRICE.define()
                     .column()
                     .nullable(false)
+                    .minimumValue(0)
                     .maximumFractionDigits(2),
             Track.RANDOM.define()
                     .column()