+++ demos/chinook/src/main/java/is/codion/framework/demos/chinook/ui/AlbumEditPanel.java
@@ -34,7 +34,7 @@ public final class AlbumEditPanel extends EntityEditPanel {
 
   public AlbumEditPanel(SwingEntityEditModel editModel) {
     super(editModel);
-    setDefaultTextFieldColumns(15);
+    defaultTextFieldColumns().set(15);
   }
 
   @Override
@@ -45,7 +45,7 @@ public final class AlbumEditPanel extends EntityEditPanel {
             .add(true)
             .edit(true);
     createTextField(Album.TITLE);
-    setComponent(Album.COVER, new CoverArtPanel(editModel().value(Album.COVER)));
+    component(Album.COVER).set(new CoverArtPanel(editModel().value(Album.COVER)));
 
     JPanel centerPanel = borderLayoutPanel()
             .westComponent(borderLayoutPanel()