+++ demos/chinook/src/main/java/is/codion/framework/demos/chinook/ui/AlbumPanel.java
@@ -28,7 +28,7 @@ public final class AlbumPanel extends EntityPanel {
     super(albumModel, new AlbumEditPanel(albumModel.editModel()), new AlbumTablePanel(albumModel.tableModel()));
     SwingEntityModel trackModel = albumModel.detailModel(Track.TYPE);
     EntityPanel trackPanel = new EntityPanel(trackModel,
-            new TrackEditPanel(trackModel.editModel()),
+            new TrackEditPanel(trackModel.editModel(), trackModel.tableModel()),
             new TrackTablePanel(trackModel.tableModel()));
 
     addDetailPanel(trackPanel);