+++ demos/chinook/src/main/java/is/codion/framework/demos/chinook/ui/TrackEditPanel.java
@@ -20,7 +20,10 @@ package is.codion.framework.demos.chinook.ui;
 
 import is.codion.framework.demos.chinook.ui.MinutesSecondsPanelValue.MinutesSecondsPanel;
 import is.codion.swing.common.ui.component.value.ComponentValue;
+import is.codion.swing.common.ui.control.Control;
+import is.codion.swing.common.ui.key.KeyEvents;
 import is.codion.swing.framework.model.SwingEntityEditModel;
+import is.codion.swing.framework.model.SwingEntityTableModel;
 import is.codion.swing.framework.ui.EntityEditPanel;
 
 import javax.swing.JPanel;
@@ -29,12 +32,19 @@ import static is.codion.framework.demos.chinook.domain.Chinook.*;
 import static is.codion.swing.common.ui.component.Components.borderLayoutPanel;
 import static is.codion.swing.common.ui.component.Components.flexibleGridLayoutPanel;
 import static is.codion.swing.common.ui.layout.Layouts.flexibleGridLayout;
+import static java.awt.event.InputEvent.CTRL_DOWN_MASK;
+import static java.awt.event.KeyEvent.VK_DOWN;
+import static java.awt.event.KeyEvent.VK_UP;
 
 public final class TrackEditPanel extends EntityEditPanel {
 
-  public TrackEditPanel(SwingEntityEditModel editModel) {
+  private final SwingEntityTableModel tableModel;
+
+  public TrackEditPanel(SwingEntityEditModel editModel, SwingEntityTableModel tableModel) {
     super(editModel);
-    setDefaultTextFieldColumns(12);
+    this.tableModel = tableModel;
+    defaultTextFieldColumns().set(12);
+    addKeyEvents();
   }
 
   @Override
@@ -93,5 +103,43 @@ public final class TrackEditPanel extends EntityEditPanel {
 
   private GenreEditPanel createGenreEditPanel() {
     return new GenreEditPanel(new SwingEntityEditModel(Genre.TYPE, editModel().connectionProvider()));
+  }
+
+  private void addKeyEvents() {
+    KeyEvents.Builder keyEvent = KeyEvents.builder()
+            .condition(WHEN_ANCESTOR_OF_FOCUSED_COMPONENT)
+            .modifiers(CTRL_DOWN_MASK);
+    keyEvent.keyCode(VK_UP)
+            .action(Control.control(this::moveSelectionUp))
+            .enable(this);
+    keyEvent.keyCode(VK_DOWN)
+            .action(Control.control(this::moveSelectionDown))
+            .enable(this);
+  }
+
+  private void moveSelectionUp() {
+    if (readyForSelectionChange()) {
+      tableModel.selectionModel().moveSelectionUp();
+    }
+  }
+
+  private void moveSelectionDown() {
+    if (readyForSelectionChange()) {
+      tableModel.selectionModel().moveSelectionDown();
+    }
+  }
+
+  private boolean readyForSelectionChange() {
+    // If the selection is empty
+    if (tableModel.selectionModel().isSelectionEmpty()) {
+      return true;
+    }
+    // If the current item is not modified
+    if (!editModel().modified().get()) {
+      return true;
+    }
+    // If the current item was modified and
+    // successfully updated after user confirmation
+    return updateWithConfirmation();
   }
 }