+++ swing/framework-ui/src/main/java/is/codion/swing/framework/ui/EntityEditComponentPanel.java
@@ -20,6 +20,7 @@ package is.codion.swing.framework.ui;
 
 import is.codion.common.Configuration;
 import is.codion.common.property.PropertyValue;
+import is.codion.common.state.State;
 import is.codion.common.value.Value;
 import is.codion.framework.domain.entity.Entities;
 import is.codion.framework.domain.entity.Entity;
@@ -107,15 +108,15 @@ public class EntityEditComponentPanel extends JPanel {
    * Default value: true
    * @see #MODIFIED_INDICATOR_UNDERLINE_STYLE
    */
-  public static final PropertyValue<Boolean> USE_MODIFIED_INDICATOR =
-          Configuration.booleanValue("is.codion.swing.framework.ui.EntityEditComponentPanel.useModifiedIndicator", true);
+  public static final PropertyValue<Boolean> MODIFIED_INDICATOR =
+          Configuration.booleanValue("is.codion.swing.framework.ui.EntityEditComponentPanel.modifiedIndicator", true);
 
   /**
    * The type of underline to use to indicate a modified value<br>
    * Value type: Integer<br>
    * Default value: {@link TextAttribute#UNDERLINE_LOW_DOTTED}<br>
    * Valid values: {@link TextAttribute}.UNDERLINE_*
-   * @see #USE_MODIFIED_INDICATOR
+   * @see #MODIFIED_INDICATOR
    */
   public static final PropertyValue<Integer> MODIFIED_INDICATOR_UNDERLINE_STYLE =
           Configuration.integerValue("is.codion.swing.framework.ui.EntityEditComponentPanel.modifiedIndicatorUnderlineStyle", TextAttribute.UNDERLINE_LOW_DOTTED);
@@ -130,7 +131,7 @@ public class EntityEditComponentPanel extends JPanel {
 
   private final SwingEntityEditModel editModel;
   private final EntityComponents entityComponents;
-  private final Map<Attribute<?>, JComponent> components = new HashMap<>();
+  private final Map<Attribute<?>, Value<JComponent>> components = new HashMap<>();
   private final Map<Attribute<?>, ComponentBuilder<?, ?, ?>> componentBuilders = new HashMap<>();
   private final Set<Attribute<?>> excludeFromSelection = new HashSet<>();
   private final Value<JComponent> focusedInputComponent = Value.value();
@@ -140,9 +141,9 @@ public class EntityEditComponentPanel extends JPanel {
   private final Value<JComponent> afterInsertFocusComponent = Value.value();
   private final Value<Attribute<?>> afterInsertFocusAttribute = Value.value();
 
-  private boolean transferFocusOnEnter = true;
-  private int defaultTextFieldColumns = DEFAULT_TEXT_FIELD_COLUMNS.get();
-  private boolean useModifiedIndicator = USE_MODIFIED_INDICATOR.get();
+  private final State transferFocusOnEnter = State.state(true);
+  private final Value<Integer> defaultTextFieldColumns = Value.value(DEFAULT_TEXT_FIELD_COLUMNS.get(), DEFAULT_TEXT_FIELD_COLUMNS.get());
+  private final State modifiedIndicator = State.state(MODIFIED_INDICATOR.get());
 
   /**
    * Instantiates a new EntityEditComponentPanel
@@ -187,16 +188,15 @@ public class EntityEditComponentPanel extends JPanel {
 
   /**
    * @param attribute the attribute
-   * @return the component associated with the given attribute
-   * @throws IllegalArgumentException in case no component or component builder has been associated with the given attribute
+   * @return the Value containing the component associated with the given attribute
    */
-  public final JComponent component(Attribute<?> attribute) {
-    JComponent component = getComponentInternal(attribute);
-    if (component == null) {
-      throw new IllegalArgumentException("No component associated with attribute: " + attribute);
+  public final Value<JComponent> component(Attribute<?> attribute) {
+    ComponentBuilder<?, ?, ?> componentBuilder = componentBuilders.get(requireNonNull(attribute));
+    if (componentBuilder != null) {
+      componentBuilder.build();
     }
 
-    return component;
+    return components.computeIfAbsent(attribute, k -> Value.value());
   }
 
   /**
@@ -275,10 +275,7 @@ public class EntityEditComponentPanel extends JPanel {
    * @param attribute the attribute of the component to select
    */
   public final void requestComponentFocus(Attribute<?> attribute) {
-    JComponent component = getComponentInternal(attribute);
-    if (component != null) {
-      focusableComponent(component).requestFocus();
-    }
+    component(attribute).optional().ifPresent(component -> focusableComponent(component).requestFocus());
   }
 
   /**
@@ -309,12 +306,12 @@ public class EntityEditComponentPanel extends JPanel {
    * this returns all (non-excluded) attributes that have an associated component in this panel
    * that is enabled, displayable, visible and focusable.
    * @see #excludeComponentsFromSelection(Attribute[])
-   * @see #setComponent(Attribute, JComponent)
+   * @see #component(Attribute)
    */
   public final Collection<Attribute<?>> selectComponentAttributes() {
     return components.keySet().stream()
             .filter(attribute -> !excludeFromSelection.contains(attribute))
-            .filter(attribute -> componentSelectable(getComponentInternal(attribute)))
+            .filter(attribute -> componentSelectable(component(attribute).get()))
             .collect(collectingAndThen(Collectors.toList(), Collections::unmodifiableCollection));
   }
 
@@ -353,49 +350,37 @@ public class EntityEditComponentPanel extends JPanel {
   /**
    * If set to true then component labels will indicate that the value is modified.
    * This applies to all components created by this edit component panel as well as
-   * components set via {@link #setComponent(Attribute, JComponent)} as long
+   * components set via {@link #component(Attribute)} as long
    * as the component has a JLabel associated with its 'labeledBy' client property.
    * Note that this has no effect on components that have already been created.
-   * @param useModifiedIndicator the new value
-   * @see #USE_MODIFIED_INDICATOR
+   * @return the State controlling whether components display an indicator if the value is modified
+   * @see #MODIFIED_INDICATOR
    * @see JLabel#setLabelFor(Component)
    */
-  protected final void setUseModifiedIndicator(boolean useModifiedIndicator) {
-    this.useModifiedIndicator = useModifiedIndicator;
+  protected final State modifiedIndicator() {
+    return modifiedIndicator;
   }
 
   /**
    * If set to true then components created subsequently will transfer focus on enter, otherwise not.
    * Note that this has no effect on components that have already been created.
-   * @param transferFocusOnEnter the new value
+   * @return the State controlling whether components transfer focus on enter
    * @see ComponentBuilder#TRANSFER_FOCUS_ON_ENTER
    */
-  protected final void setTransferFocusOnEnter(boolean transferFocusOnEnter) {
-    this.transferFocusOnEnter = transferFocusOnEnter;
+  protected final State transferFocusOnEnter() {
+    return transferFocusOnEnter;
   }
 
   /**
-   * Sets the default number of text field columns, -1 for not settings the columns
-   * @param defaultTextFieldColumns the default text field columns
+   * Controls the default number of text field columns, -1 for not settings the columns
+   * @return the Value controlling the default number of text field columns
    * @see #DEFAULT_TEXT_FIELD_COLUMNS
    * @see #createTextField(Attribute)
    * @see #createForeignKeySearchField(ForeignKey)
    * @see #createTextFieldPanel(Attribute)
    */
-  protected final void setDefaultTextFieldColumns(int defaultTextFieldColumns) {
-    this.defaultTextFieldColumns = defaultTextFieldColumns;
-  }
-
-  /**
-   * Associates the given input component with the given attribute.
-   * @param attribute the attribute
-   * @param component the input component
-   */
-  protected final void setComponent(Attribute<?> attribute, JComponent component) {
-    components.put(requireNonNull(attribute), requireNonNull(component));
-    if (useModifiedIndicator && attribute.entityType().equals(editModel.entityType())) {
-      editModel.modified(attribute).addDataListener(new ModifiedIndicator(component));
-    }
+  protected final Value<Integer> defaultTextFieldColumns() {
+    return defaultTextFieldColumns;
   }
 
   /**
@@ -447,7 +432,7 @@ public class EntityEditComponentPanel extends JPanel {
    * @throws IllegalArgumentException in case no component has been associated with the given attribute
    */
   protected final JPanel createInputPanel(Attribute<?> attribute) {
-    return createInputPanel(attribute, component(attribute));
+    return createInputPanel(attribute, getComponentOrThrow(attribute));
   }
 
   /**
@@ -520,7 +505,8 @@ public class EntityEditComponentPanel extends JPanel {
       // component, f.ex. this could be a text area wrapped in a scroll pane or a combobox on a panel with
       // a new instance button, we assume that this input component at least contains the actual component
       components.values().stream()
-              .filter(comp -> sameOrParentOf(inputComponent, comp))
+              .map(Value::get)
+              .filter(component -> sameOrParentOf(inputComponent, component))
               .findAny()
               .ifPresent(component -> setLabelForComponent((JLabel) labelComponent, component));
     }
@@ -547,7 +533,7 @@ public class EntityEditComponentPanel extends JPanel {
    */
   protected final TextFieldPanel.Builder createTextFieldPanel(Attribute<String> attribute) {
     return setComponentBuilder(attribute, entityComponents.textFieldPanel(attribute)
-            .columns(defaultTextFieldColumns)
+            .columns(defaultTextFieldColumns.get())
             .onBuild(inputPanel -> addValidator(attribute, inputPanel.textField(), editModel())));
   }
 
@@ -584,7 +570,7 @@ public class EntityEditComponentPanel extends JPanel {
    */
   protected final <T, C extends JTextField, B extends TextFieldBuilder<T, C, B>> TextFieldBuilder<T, C, B> createTextField(Attribute<T> attribute) {
     return setComponentBuilder(attribute, (TextFieldBuilder<T, C, B>) entityComponents.textField(attribute)
-            .columns(defaultTextFieldColumns)
+            .columns(defaultTextFieldColumns.get())
             .onBuild(field -> addFormattedValidator(attribute, field, editModel())));
   }
 
@@ -666,7 +652,7 @@ public class EntityEditComponentPanel extends JPanel {
    */
   protected final NumberField.Builder<Integer> createIntegerField(Attribute<Integer> attribute) {
     return setComponentBuilder(attribute, entityComponents.integerField(attribute)
-            .columns(defaultTextFieldColumns)
+            .columns(defaultTextFieldColumns.get())
             .onBuild(field -> addValidator(attribute, field, editModel())));
   }
 
@@ -677,7 +663,7 @@ public class EntityEditComponentPanel extends JPanel {
    */
   protected final NumberField.Builder<Long> createLongField(Attribute<Long> attribute) {
     return setComponentBuilder(attribute, entityComponents.longField(attribute)
-            .columns(defaultTextFieldColumns)
+            .columns(defaultTextFieldColumns.get())
             .onBuild(field -> addValidator(attribute, field, editModel())));
   }
 
@@ -688,7 +674,7 @@ public class EntityEditComponentPanel extends JPanel {
    */
   protected final NumberField.Builder<Double> createDoubleField(Attribute<Double> attribute) {
     return setComponentBuilder(attribute, entityComponents.doubleField(attribute)
-            .columns(defaultTextFieldColumns)
+            .columns(defaultTextFieldColumns.get())
             .onBuild(field -> addValidator(attribute, field, editModel())));
   }
 
@@ -699,7 +685,7 @@ public class EntityEditComponentPanel extends JPanel {
    */
   protected final NumberField.Builder<BigDecimal> createBigDecimalField(Attribute<BigDecimal> attribute) {
     return setComponentBuilder(attribute, entityComponents.bigDecimalField(attribute)
-            .columns(defaultTextFieldColumns)
+            .columns(defaultTextFieldColumns.get())
             .onBuild(field -> addValidator(attribute, field, editModel())));
   }
 
@@ -808,7 +794,7 @@ public class EntityEditComponentPanel extends JPanel {
   protected final EntitySearchField.Builder createForeignKeySearchField(ForeignKey foreignKey) {
     return setComponentBuilder(foreignKey, entityComponents.foreignKeySearchField(foreignKey,
                     editModel().foreignKeySearchModel(foreignKey))
-            .columns(defaultTextFieldColumns));
+            .columns(defaultTextFieldColumns.get()));
   }
 
   /**
@@ -821,7 +807,7 @@ public class EntityEditComponentPanel extends JPanel {
                                                                                   Supplier<EntityEditPanel> editPanelSupplier) {
     return setComponentBuilder(foreignKey, entityComponents.foreignKeySearchFieldPanel(foreignKey,
                     editModel().foreignKeySearchModel(foreignKey), editPanelSupplier)
-            .columns(defaultTextFieldColumns));
+            .columns(defaultTextFieldColumns.get()));
   }
 
   /**
@@ -855,7 +841,7 @@ public class EntityEditComponentPanel extends JPanel {
             .definition(requireNonNull(attribute).entityType()).attributes().definition(attribute);
     return (LabelBuilder<T>) Components.label(attributeDefinition.caption())
             .displayedMnemonic(attributeDefinition.mnemonic())
-            .labelFor(getComponentInternal(attribute));
+            .labelFor(component(attribute).get());
   }
 
   /**
@@ -867,7 +853,7 @@ public class EntityEditComponentPanel extends JPanel {
     }
 
     if (initialFocusAttribute.isNotNull()) {
-      return getComponentInternal(initialFocusAttribute.get());
+      return component(initialFocusAttribute.get()).get();
     }
 
     return null;
@@ -882,7 +868,7 @@ public class EntityEditComponentPanel extends JPanel {
     }
 
     if (afterInsertFocusAttribute.isNotNull()) {
-      return getComponentInternal(afterInsertFocusAttribute.get());
+      return component(afterInsertFocusAttribute.get()).get();
     }
 
     return getInitialFocusComponent();
@@ -903,7 +889,7 @@ public class EntityEditComponentPanel extends JPanel {
       throw new IllegalStateException("ComponentBuilder has already been set for attribute: " + attribute);
     }
     componentBuilders.put(attribute, componentBuilder
-            .transferFocusOnEnter(transferFocusOnEnter)
+            .transferFocusOnEnter(transferFocusOnEnter.get())
             .linkedValue(editModel().value(attribute))
             .onBuild(new OnComponentBuilt<>(attribute)));
 
@@ -919,13 +905,13 @@ public class EntityEditComponentPanel extends JPanel {
     }
   }
 
-  private JComponent getComponentInternal(Attribute<?> attribute) {
-    ComponentBuilder<?, ?, ?> componentBuilder = componentBuilders.get(requireNonNull(attribute));
-    if (componentBuilder != null) {
-      componentBuilder.build();
+  private JComponent getComponentOrThrow(Attribute<?> attribute) {
+    Value<JComponent> component = component(attribute);
+    if (component.isNull()) {
+      throw new IllegalArgumentException("No component associated with attribute: " + attribute);
     }
 
-    return components.get(attribute);
+    return component.get();
   }
 
   private void addFocusedComponentListener() {
@@ -1004,6 +990,13 @@ public class EntityEditComponentPanel extends JPanel {
       componentBuilders.remove(attribute);
       setComponent(attribute, component);
     }
+
+    private void setComponent(Attribute<?> attribute, JComponent component) {
+      components.computeIfAbsent(requireNonNull(attribute), k -> Value.value()).set(requireNonNull(component));
+      if (modifiedIndicator.get() && attribute.entityType().equals(editModel.entityType())) {
+        editModel.modified(attribute).addDataListener(new ModifiedIndicator(component));
+      }
+    }
   }
 
   private final class FocusedInputComponentListener implements PropertyChangeListener {
@@ -1017,13 +1010,10 @@ public class EntityEditComponentPanel extends JPanel {
     }
 
     private boolean inputComponent(JComponent component) {
-      for (JComponent inputComponent : components.values()) {
-        if (sameOrParentOf(inputComponent, component)) {
-          return true;
-        }
-      }
-
-      return false;
+      return components.values().stream()
+              .filter(Value::isNotNull)
+              .map(Value::get)
+              .anyMatch(comp -> sameOrParentOf(comp, component));
     }
   }